Use uint16 in winsize_unsupported.go (#136)

+++ winsize_unsupported.go
@@ -9,7 +9,7 @@ import (
 
 // Winsize is a dummy struct to enable compilation on unsupported platforms.
 type Winsize struct {
-	Rows, Cols, X, Y uint
+	Rows, Cols, X, Y uint16
 }
 
 // Setsize resizes t to s.
